Extract error messages into constant
